refactor) postCardUi top,content,bottom 파일 분리
