+++ index.html
@@ -69,7 +69,7 @@
             <div id="postCards"> </div>
 
 
-            
+
 
             <!--Figure3에만 등장하는 부분-->
             <!--하단 footer(navbar) 부분-->
@@ -115,7 +115,7 @@
     <script src="js/components/navBarUi.js"></script>
     <script src="js/components/storyUi.js"></script>
     <script src="js/components/storyFeat.js"></script>
-    <script src="js/components/postCardUi.js"></script>
+    <script src="js/components/postCard/Ui/postCardUi.js" type="module"></script>
 </body>
 
 </html>
